+++ lib/smart_core/types.rb
@@ -10,4 +10,5 @@ module SmartCore::Types
   require_relative 'types/system'
   require_relative 'types/primitive'
   require_relative 'types/value'
+  require_relative 'types/protocol'
 end