+++ lib/smart_core/types/primitive/caster.rb
@@ -2,6 +2,7 @@
 
 # @api private
 # @since 0.1.0
+# @version 0.3.0
 class SmartCore::Types::Primitive::Caster
   # @param expression [Proc]
   # @return [void]
@@ -13,12 +14,14 @@ class SmartCore::Types::Primitive::Caster
   end
 
   # @param value [Any]
+  # @param runtime_attributes [Array<Any>]
   # @return [Any]
   #
   # @api private
   # @since 0.1.0
-  def call(value)
-    expression.call(value)
+  # @version 0.3.0
+  def call(value, runtime_attributes)
+    expression.call(value, runtime_attributes)
   end
 
   private