+++ lib/smart_core/types/primitive/invariant_control/chain.rb
@@ -31,14 +31,15 @@ class SmartCore::Types::Primitive::InvariantControl::Chain
   end
 
   # @param value [Any]
+  # @param runtime_attributes [Array<Any>]
   # @return [SmartCore::Types::Primitive::InvariantControl::Chain::Result]
   #
   # @api private
   # @since 0.2.0
-  def check(value)
+  def check(value, runtime_attributes)
     invariant_results = [].tap do |results|
       invariants.each do |invariant|
-        result = invariant.check(value).tap { |res| results << res }
+        result = invariant.check(value, runtime_attributes).tap { |res| results << res }
         break if result.failure?
       end
     end