+++ lib/smart_core/types/primitive/undefined_caster.rb
@@ -13,13 +13,14 @@ class SmartCore::Types::Primitive::UndefinedCaster < SmartCore::Types::Primitive
   end
 
   # @param value [Any]
+  # @param runtime_attributes [Array<Any>]
   # @return [void]
   #
   # @raise [SmartCore::Types::TypeCastingUnsupportedError]
   #
   # @pai private
   # @since 0.1.0
-  def call(value)
+  def call(value, runtime_attributes)
     raise(SmartCore::Types::TypeCastingUnsupportedError, <<~ERROR_MESSAGE)
       'This type has no support for type casting'
     ERROR_MESSAGE