+++ lib/smart_core/types/protocol.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# @api public
+# @since 0.3.0
+class SmartCore::Types::Protocol < SmartCore::Types::Primitive
+  require_relative 'protocol/instance_of'
+end